refactor: extract combat sidebar toggle to separate module
- Move combat toggle logic from ready.ts to combatSidebarToggle.ts
- Block right-click context menu on combat sidebar button
- Prevents native Foundry multi-encounter menu from appearing
